Make ServerPropertiesControl's comboboxes populate with the actual properties from server.properties.
- Added more converters to Converters.cs
- Created Services folder
- Added ServerPropertiesService to Services
- Added methods to parse and save information from server.properties.
- Added calls to parse and save server.properties in MainWindow's event handlers.

P.S. Making this work was such a fucking pain in the ass. At the moment only the most important fields from server.properties are in the sidebar panel, but in the future i should be able to add an advanced server properties window that handles all the fields inside server.properties.